# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SQLoptimize API",
    "description": "SQLoptimize API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/models/active": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "获取用户当前活跃的模型连接",
        "description": "获取用户当前活跃的模型连接",
        "operationId": "get_active_model_connection_api_models_active_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveModelConnectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "models"
        ],
        "summary": "设置用户当前活跃的模型连接",
        "description": "设置用户当前活跃的模型连接",
        "operationId": "set_active_model_connection_api_models_active_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetActiveModelConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetActiveModelConnectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/models/": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "获取用户模型连接列表",
        "description": "获取当前用户的模型连接列表",
        "operationId": "get_user_models_api_models__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConnectionListResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "models"
        ],
        "summary": "创建模型连接",
        "description": "创建新的模型连接",
        "operationId": "create_model_connection_api_models__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelConnectionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConnectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{connection_id}": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "根据ID获取模型连接",
        "description": "获取指定的模型连接信息",
        "operationId": "get_model_connection_api_models__connection_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConnectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "models"
        ],
        "summary": "更新模型连接",
        "description": "更新模型连接信息",
        "operationId": "update_model_connection_api_models__connection_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Connection Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelConnectionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConnectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "models"
        ],
        "summary": "删除模型连接",
        "description": "删除模型连接",
        "operationId": "delete_model_connection_api_models__connection_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActiveModelConnectionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "模型连接ID"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "description": "用户自定义的模型名称"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "模型名称"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url",
            "description": "模型API地址"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "description": "混淆后的模型API密钥"
          },
          "model_description": {
            "type": "string",
            "title": "Model Description",
            "description": "模型描述"
          },
          "model_avatar_url": {
            "type": "string",
            "title": "Model Avatar Url",
            "description": "模型头像URL"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "更新时间"
          }
        },
        "type": "object",
        "required": [
          "id",
          "model_name",
          "model",
          "base_url",
          "api_key",
          "model_description",
          "model_avatar_url",
          "created_at",
          "updated_at"
        ],
        "title": "ActiveModelConnectionResponse",
        "description": "活跃模型连接响应"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelConnectionCreate": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "description": "用户自定义的模型名称"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "模型名称"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url",
            "description": "模型API地址"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "description": "模型API密钥"
          },
          "model_description": {
            "type": "string",
            "title": "Model Description",
            "description": "模型描述",
            "default": ""
          },
          "model_avatar_url": {
            "type": "string",
            "title": "Model Avatar Url",
            "description": "模型头像URL",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "model_name",
          "model",
          "base_url",
          "api_key"
        ],
        "title": "ModelConnectionCreate",
        "description": "创建模型连接请求"
      },
      "ModelConnectionListResponse": {
        "properties": {
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelConnectionResponse"
            },
            "type": "array",
            "title": "Models",
            "description": "模型连接列表"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "总数"
          },
          "skip": {
            "type": "integer",
            "title": "Skip",
            "description": "跳过数量"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "限制数量"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "是否还有更多数据"
          },
          "active_connection_id": {
            "type": "integer",
            "title": "Active Connection Id",
            "description": "当前用户活跃的模型连接ID，0表示无活跃连接",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "models",
          "total",
          "skip",
          "limit",
          "has_more"
        ],
        "title": "ModelConnectionListResponse",
        "description": "模型连接列表响应"
      },
      "ModelConnectionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "模型连接ID"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "description": "用户自定义的模型名称"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "模型名称"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url",
            "description": "模型API地址"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "description": "混淆后的模型API密钥"
          },
          "model_description": {
            "type": "string",
            "title": "Model Description",
            "description": "模型描述"
          },
          "model_avatar_url": {
            "type": "string",
            "title": "Model Avatar Url",
            "description": "模型头像URL"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "更新时间"
          }
        },
        "type": "object",
        "required": [
          "id",
          "model_name",
          "model",
          "base_url",
          "api_key",
          "model_description",
          "model_avatar_url",
          "created_at",
          "updated_at"
        ],
        "title": "ModelConnectionResponse",
        "description": "模型连接响应"
      },
      "ModelConnectionUpdate": {
        "properties": {
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name",
            "description": "用户自定义的模型名称"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "模型名称"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url",
            "description": "模型API地址"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key",
            "description": "模型API密钥"
          },
          "model_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Description",
            "description": "模型描述"
          },
          "model_avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Avatar Url",
            "description": "模型头像URL"
          }
        },
        "type": "object",
        "title": "ModelConnectionUpdate",
        "description": "更新模型连接请求"
      },
...
      "SetActiveModelConnectionRequest": {
        "properties": {
          "connection_id": {
            "type": "integer",
            "title": "Connection Id",
            "description": "模型连接ID"
          }
        },
        "type": "object",
        "required": [
          "connection_id"
        ],
        "title": "SetActiveModelConnectionRequest",
        "description": "设置活跃模型连接请求"
      },
      "SetActiveModelConnectionResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "设置结果消息"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "SetActiveModelConnectionResponse",
        "description": "设置活跃模型连接响应"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}